+++ .github/scripts/provision-linux-test.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 set -euo pipefail
-sudo apt-get update && sudo apt-get install -y softhsm2
+sudo apt-get update && sudo apt-get install -y softhsm2 mitmproxy